Turbopack: remove ChunkableModuleReference
