+++ crates/next-core/src/next_server_component/server_component_reference.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use turbo_rcstr::RcStr;
 use turbo_tasks::{ResolvedVc, ValueToString, Vc};
 use turbopack_core::{
-    chunk::{ChunkableModuleReference, ChunkingType, ChunkingTypeOption},
+    chunk::{ChunkingType, ChunkingTypeOption},
     module::Module,
     reference::ModuleReference,
     resolve::ModuleResolveResult,
@@ -41,10 +41,6 @@ impl ModuleReference for NextServerComponentModuleReference {
     fn resolve_reference(&self) -> Vc<ModuleResolveResult> {
         *ModuleResolveResult::module(self.asset)
     }
-}
-
-#[turbo_tasks::value_impl]
-impl ChunkableModuleReference for NextServerComponentModuleReference {
     #[turbo_tasks::function]
     fn chunking_type(&self) -> Vc<ChunkingTypeOption> {
         Vc::cell(Some(ChunkingType::Shared {
