+++ turbopack/crates/turbopack-core/src/introspect/utils.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use turbo_rcstr::{RcStr, rcstr};
-use turbo_tasks::{FxIndexSet, ResolvedVc, Vc};
+use turbo_tasks::{FxIndexSet, Vc};
 use turbo_tasks_fs::FileContent;
 
 use super::{
@@ -8,7 +8,7 @@ use super::{
 };
 use crate::{
     asset::AssetContent,
-    chunk::{ChunkableModuleReference, ChunkingType},
+    chunk::ChunkingType,
     output::OutputAssetsWithReferenced,
     reference::{ModuleReference, ModuleReferences},
 };
@@ -68,25 +68,19 @@ pub async fn children_from_module_references(
     let mut children = FxIndexSet::default();
     let references = references.await?;
     for &reference in &*references {
-        let key = if let Some(chunkable) =
-            ResolvedVc::try_downcast::<Box<dyn ChunkableModuleReference>>(reference)
-        {
-            match &*chunkable.chunking_type().await? {
-                None => key.clone(),
-                Some(ChunkingType::Parallel { inherit_async, .. }) => {
-                    if *inherit_async {
-                        parallel_inherit_async_reference_ty()
-                    } else {
-                        parallel_reference_ty()
-                    }
+        let key = match &*reference.chunking_type().await? {
+            None => key.clone(),
+            Some(ChunkingType::Parallel { inherit_async, .. }) => {
+                if *inherit_async {
+                    parallel_inherit_async_reference_ty()
+                } else {
+                    parallel_reference_ty()
                 }
-                Some(ChunkingType::Async) => async_reference_ty(),
-                Some(ChunkingType::Isolated { .. }) => isolated_reference_ty(),
-                Some(ChunkingType::Shared { .. }) => shared_reference_ty(),
-                Some(ChunkingType::Traced) => traced_reference_ty(),
             }
-        } else {
-            key.clone()
+            Some(ChunkingType::Async) => async_reference_ty(),
+            Some(ChunkingType::Isolated { .. }) => isolated_reference_ty(),
+            Some(ChunkingType::Shared { .. }) => shared_reference_ty(),
+            Some(ChunkingType::Traced) => traced_reference_ty(),
         };
 
         for &module in reference
