+++ turbopack/crates/turbopack-ecmascript/src/references/worker.rs
@@ -11,7 +11,7 @@ use turbo_tasks::{
 };
 use turbo_tasks_fs::FileSystemPath;
 use turbopack_core::{
-    chunk::{ChunkableModule, ChunkableModuleReference, ChunkingContext, EvaluatableAsset},
+    chunk::{ChunkableModule, ChunkingContext, ChunkingType, ChunkingTypeOption, EvaluatableAsset},
     context::AssetContext,
     issue::{IssueExt, IssueSeverity, IssueSource, StyledString, code_gen::CodeGenerationIssue},
     module::Module,
@@ -255,6 +255,14 @@ impl ModuleReference for WorkerAssetReference {
         }
         .cell())
     }
+
+    #[turbo_tasks::function]
+    fn chunking_type(self: Vc<Self>) -> Vc<ChunkingTypeOption> {
+        Vc::cell(Some(ChunkingType::Parallel {
+            inherit_async: false,
+            hoisted: false,
+        }))
+    }
 }
 
 impl WorkerAssetReference {
@@ -293,9 +301,6 @@ impl ValueToString for WorkerAssetReference {
         ))
     }
 }
-
-#[turbo_tasks::value_impl]
-impl ChunkableModuleReference for WorkerAssetReference {}
 
 impl IntoCodeGenReference for WorkerAssetReference {
     fn into_code_gen_reference(
